refactor that huge ComputeJobs index component
